requirements.txt
# Core Azure AI and OpenAI packages
azure-ai-projects>=1.0.0b12
azure-ai-agents>=1.1.0b4  # Pre-release version needed for DeepResearchTool
azure-ai-evaluation
azure-identity>=1.23.1
azure-search-documents>=11.4.0
python-dotenv>=1.1.1
openai>=1.0.0

# LangChain for LangChain samples
langchain-openai>=0.1.0

# Data analysis and visualization (for Extraction notebook)
pandas>=2.0.0
matplotlib>=3.7.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and testing tools
pytest>=8.0.0
flake8>=7.0.0
ruff>=0.12.0

# Additional useful packages
requests>=2.32.0
numpy>=1.24.0

# GUI dependencies
markdown>=3.4.0
tkhtmlview>=0.3.0

# PDF generation
reportlab>=4.0.0

# Azure AI Foundry tracing and monitoring
azure-monitor-opentelemetry>=1.2.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-openai-v2

# Azure AI Evaluation
azure-ai-evaluation>=1.10.0

# For Playwright MCP integration (MCP tool support should be in azure-ai-agents)
# Note: Playwright MCP server runs as separate Node.js process
# No additional Python packages needed for MCP client functionality
